python_server/requirements.txt

flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.1
matplotlib==3.7.2
numpy==1.24.3
